fix: divided can comm api
