+++ Core/Inc/bms/post/post.h
@@ -13,7 +13,7 @@
 
 #include "cellboard-def.h"
 
-#include "can-comm.h"
+#include "can-comm-api.h"
 #include "bms-manager.h"
 #include "led.h"
 #include "temp.h"
