+++ platformio.ini
@@ -46,6 +46,7 @@ build_flags =
     -ICore/Inc/bms/balancing
     -ICore/Inc/bms/post
     -ICore/Inc/bms/temp
+    -ICore/Inc/bms/can-comm
     -ICore/Lib
     -ICore/Lib/can/lib
     -ICore/Lib/can/lib/bms
@@ -120,6 +121,7 @@ build_flags =
     -ICore/Inc/bms/balancing
     -ICore/Inc/bms/post
     -ICore/Inc/bms/temp
+    -ICore/Inc/bms/can-comm
     -ICore/Lib
     -ICore/Lib/can/lib
     -ICore/Lib/can/lib/bms
